Added Logging library
Added logging library from loguru : https://github.com/emilk/loguru as a submodule so that we can update this easily. We should bundle it for releases.

Updated the existing skeleton classes to give some examples of the logger.

Updated README to give example of how to clone project for development work.